cleanup, parsed to components
