+++ components/catalogue/categoryList/CategoryList.cont.tsx
@@ -1,11 +1,9 @@
-import ModalComponent from 'components/ui/modal/modal.comp'
 import { useCategoryPath } from 'hooks/usePath'
 import { useSession } from 'next-auth/react'
-import { Dispatch, Fragment, SetStateAction, useEffect, useState } from 'react'
+import { Dispatch, Fragment, SetStateAction, useEffect } from 'react'
 import useSWR from 'swr'
 import { CatalogueCategoryResponse } from 'types/responses'
 
-import TestEditModal from '../categoryEditForm/TestEdit'
 import CategoryItemComponent from './CategoryItem.comp'
 
 interface Props {
@@ -14,7 +12,6 @@ interface Props {
 
 const CategoryListContainer = ({ setCatalogueCategoryList }: Props) => {
   const { data: session } = useSession()
-  const [open, setopen] = useState(false)
   const categoryPath = useCategoryPath()
   /* fetch category list */
   const { data: categoryList } = useSWR<Array<CatalogueCategoryResponse>>(session ? categoryPath : null)
@@ -32,28 +29,9 @@ const CategoryListContainer = ({ setCatalogueCategoryList }: Props) => {
             {categoryList?.map(category => (
               <CategoryItemComponent key={category.code} category={category} />
             ))}
-            {/* <button
-              onClick={() => {
-                setopen(true)
-              }}
-              className="relative flex items-center space-x-3 rounded-lg border border-gray-300 bg-white px-6 py-5 shadow-sm focus-within:ring-2 focus-within:ring-primary-500 focus-within:ring-offset-2 hover:border-gray-400"
-            >
-              <div className="flex-shrink-0">
-                <PlusIcon className="h-10 w-10 rounded-full" />
-              </div>
-              <div className="min-w-0 flex-1">
-                <div className="focus:outline-none">
-                  <span className="absolute inset-0" aria-hidden="true" />
-                  <p className="text-sm font-medium text-gray-900">New Category</p>
-                </div>
-              </div>
-            </button> */}
           </div>
         </div>
       )}
-      <ModalComponent open={open} setOpen={setopen} buttons={{ noButtons: true }} testid="catalogueEdit">
-        <TestEditModal setopen={setopen} />
-      </ModalComponent>
     </Fragment>
   )
 }
